feat: Change menu structure.
