+++ src/router/modules/ADempiere/menu.js
@@ -25,7 +25,7 @@ import staticRoutes from '@/router/modules/ADempiere/staticRoutes.js'
 import { requestMenu } from '@/api/ADempiere/security/index.ts'
 
 // Utils and Helper Methods
-import { convertAction } from '@/utils/ADempiere/dictionaryUtils.js'
+import { convertAction, getReferenceByMenu } from '@/utils/ADempiere/dictionary/menu'
 import { getCurrentClient, getCurrentOrganization, getCurrentRole } from '@/utils/ADempiere/auth'
 
 /**
@@ -83,7 +83,7 @@ export function loadMainMenu({
         asyncRoutesMap.push(optionMenu)
       })
 
-      const permiseStactiRoutes = hidenStactiRoutes({
+      const permiseStactiRoutes = hidenStaticRoutes({
         staticRoutes,
         permiseRole: role
       })
@@ -109,7 +109,9 @@ export function loadMainMenu({
  */
 function getChildFromAction({ menu, index, clientId, roleId, organizationId }) {
   const { component, icon, name: type } = convertAction(menu.action)
-  const routeIdentifier = type + '/' + menu.reference_id
+  const { reference_id, reference_uuid } = getReferenceByMenu(menu)
+
+  const routeIdentifier = type + '/' + reference_id
   const isIndex = menu.is_summary
   const option = {
     path: '/' + clientId + '/' + roleId + '/' + organizationId + '/' + menu.id + '/' + routeIdentifier,
@@ -125,16 +127,16 @@ function getChildFromAction({ menu, index, clientId, roleId, organizationId }) {
       isSummary: menu.is_summary,
       isSalesTransaction: menu.is_sales_transaction,
       parentId: menu.parent_id,
-      parentUuid: menu.parent_uuid,
+      // parentUuid: menu.parent_uuid,
       noCache: false,
-      referenceId: menu.reference_id,
-      referenceUuid: menu.reference_uuid,
       tabUuid: '',
       title: menu.name,
       type,
-      id: menu.reference_id,
-      uuid: menu.reference_uuid,
-      containerKey: type + '_' + menu.reference_id,
+      id: reference_id,
+      uuid: reference_uuid,
+      referenceId: reference_id,
+      referenceUuid: reference_uuid,
+      containerKey: type + '_' + reference_id,
       childs: []
     },
     children: []
@@ -169,6 +171,8 @@ function getChildFromAction({ menu, index, clientId, roleId, organizationId }) {
 function getRouteFromMenuItem({ menu, clientId, roleId, organizationId }) {
   // use component of convertAction
   const { icon, name: type } = convertAction(menu.action)
+  const { reference_id, reference_uuid } = getReferenceByMenu(menu)
+
   const isIndex = menu.is_summary
   const optionMenu = {
     path: '/' + clientId + '/' + roleId + '/' + organizationId + '/' + menu.id,
@@ -176,8 +180,6 @@ function getRouteFromMenuItem({ menu, clientId, roleId, organizationId }) {
     component: Layout,
     name: menu.id.toString(),
     meta: {
-      id: menu.reference_id,
-      uuid: menu.reference_uuid,
       description: menu.description,
       icon,
       isIndex,
@@ -185,11 +187,13 @@ function getRouteFromMenuItem({ menu, clientId, roleId, organizationId }) {
       isSummary: menu.is_summary,
       isSalesTransaction: menu.is_sales_transaction,
       parentId: menu.parent_id,
-      parentUuid: menu.parent_uuid,
+      // parentUuid: menu.parent_uuid,
       noCache: true,
-      referenceId: menu.reference_id,
-      referenceUuid: menu.reference_uuid,
-      containerKey: type + '_' + menu.reference_id,
+      id: reference_id,
+      uuid: reference_uuid,
+      referenceId: reference_id,
+      referenceUuid: reference_uuid,
+      containerKey: type + '_' + reference_id,
       title: menu.name,
       type,
       childs: []
@@ -206,7 +210,7 @@ function getRouteFromMenuItem({ menu, clientId, roleId, organizationId }) {
  * @param {object} permiseRole role permissions
  * @returns {object} routes with hidden/show
  */
-function hidenStactiRoutes({ staticRoutes, permiseRole }) {
+function hidenStaticRoutes({ staticRoutes, permiseRole }) {
   if (!permiseRole) {
     return staticRoutes
   }
