+++ src/utils/ADempiere/dictionary/menu.js
@@ -1,0 +1,114 @@
+/**
+ * ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
+ * Copyright (C) 2018-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ * Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+/**
+ * Convert action to action name for route
+ * @param {string} action
+ * @returns {object}
+ */
+export function convertAction(action) {
+  const actionAttributes = {
+    name: '',
+    icon: '',
+    hidden: false,
+    isIndex: false,
+    component: () => import('@/views/ADempiere/Unsupported')
+  }
+  switch (action) {
+    case 'B':
+      actionAttributes.name = 'workbech'
+      actionAttributes.icon = 'peoples'
+      break
+    case 'F':
+      actionAttributes.name = 'workflow'
+      actionAttributes.icon = 'example'
+      actionAttributes.component = () => import('@/views/ADempiere/Workflow')
+      break
+    case 'P':
+      actionAttributes.name = 'process'
+      actionAttributes.icon = 'component'
+      actionAttributes.component = () => import('@/views/ADempiere/Process')
+      break
+    case 'R':
+      actionAttributes.name = 'report'
+      actionAttributes.icon = 'skill'
+      actionAttributes.component = () => import('@/views/ADempiere/Report')
+      break
+    case 'S':
+      actionAttributes.name = 'browser'
+      actionAttributes.icon = 'search'
+      actionAttributes.component = () => import('@/views/ADempiere/Browser')
+      break
+    case 'T':
+      actionAttributes.name = 'task'
+      actionAttributes.icon = 'size'
+      break
+    case 'W':
+      actionAttributes.name = 'window'
+      actionAttributes.icon = 'tab'
+      actionAttributes.component = () => import('@/views/ADempiere/Window')
+      break
+    case 'X':
+      actionAttributes.name = 'form'
+      actionAttributes.icon = 'form'
+      actionAttributes.component = () => import('@/views/ADempiere/Form')
+      break
+    default:
+      actionAttributes.name = 'summary'
+      actionAttributes.icon = 'nested'
+      // actionAttributes.hidden = true
+      actionAttributes.isIndex = true
+      actionAttributes.component = () => import('@/views/ADempiere/Summary')
+      break
+  }
+  return actionAttributes
+}
+
+export function getReferenceByMenu(menu) {
+  let reference_id = 0
+  let reference_uuid = ''
+
+  switch (menu.action) {
+    case 'F':
+      reference_id = menu.workflow.id
+      reference_uuid = menu.workflow.uuid
+      break
+    case 'P':
+    case 'R':
+      reference_id = menu.process.id
+      reference_uuid = menu.process.uuid
+      break
+    case 'S':
+      reference_id = menu.browse.id
+      reference_uuid = menu.browse.uuid
+      break
+    case 'W':
+      reference_id = menu.window.id
+      reference_uuid = menu.window.uuid
+      break
+    case 'X':
+      reference_id = menu.form.id
+      reference_uuid = menu.form.uuid
+      break
+  }
+
+  return {
+    reference_id,
+    reference_uuid
+  }
+}
